feat: Add coordinate conversion and simplification functions
This commit adds two new functions to the preprocess.js file: convertCoordinates and simplifyDisplayTitles. The convertCoordinates function takes in data and a projection and converts the coordinates of each feature in the data to x and y values using the projection. The simplifyDisplayTitles function simplifies the titles of the features in the data based on a predefined mapping.

These functions improve the data processing and display capabilities of the application.

#changelog
- Added convertCoordinates function
- Added simplifyDisplayTitles function